+++ src/ext_widgets/mledit/mledit.c
@@ -628,6 +628,9 @@ static ret_t mledit_on_event(widget_t* widget, event_t* e) {
 #else
       bool_t is_control = evt->ctrl;
 #endif
+      if (key == TK_KEY_ESCAPE || (key >= TK_KEY_F1 && key <= TK_KEY_F24)) {
+        break;
+      }
 
       if (mledit->readonly) {
         if (is_control && (key == TK_KEY_C || key == TK_KEY_c)) {
@@ -691,11 +694,17 @@ static ret_t mledit_on_event(widget_t* widget, event_t* e) {
       break;
     }
     case EVT_KEY_UP: {
-      key_event_t* key_event = key_event_cast(e);
-      if (key_code_is_enter(key_event->key)) {
+      key_event_t* evt = key_event_cast(e);
+      int32_t key = evt->key;
+
+      if (key == TK_KEY_ESCAPE || (key >= TK_KEY_F1 && key <= TK_KEY_F24)) {
+        break;
+      }
+
+      if (key_code_is_enter(key)) {
         ret = RET_STOP;
       } else {
-        ret = text_edit_key_up(mledit->model, key_event);
+        ret = text_edit_key_up(mledit->model, evt);
       }
       mledit->last_user_action_time = e->time;
       widget_invalidate(widget, NULL);