+++ docs/changes.md
@@ -1,7 +1,8 @@
 # 最新动态
 
 2021/06/16
-* 支持[极简键盘 (3 键键盘、5 键键盘和旋转按钮）](docs/how_to_support_5keys_3keys.md)
+  * 支持[极简键盘 (3 键键盘、5 键键盘和旋转按钮）](docs/how_to_support_5keys_3keys.md)
+  * edit/mledit不处理escape和功能键。 
 
 2021/06/15
   * edit 不再用左右键切换焦点。